requirements.txt
langchain>=0.0.119
langchain-openai
faiss-cpu
python-dotenv
flask
flask-cors
gunicorn